precedents-analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
aiohttp==3.9.1
python-dotenv==1.0.0
lxml==4.9.3
